pdfplumber_test/requirements.txt
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.22
PyMuPDF==1.26.3
pypdfium2==4.30.1
